backend/requirements.txt
# Core ML packages
numpy==1.23.5
pandas==1.4.4
scikit-learn==1.1.3
xgboost==1.7.3
lightgbm==3.3.5
catboost==1.2.7

# PyTorch ecosystem
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
pytorch-lightning==1.9.5
pytorch-forecasting==0.10.3
pytorch-optimizer==1.3.1

# Web dependencies
flask==2.0.2
werkzeug==2.0.3
Flask-Cors==3.0.10

# Data and visualization
matplotlib==3.7.1
statsmodels==0.13.5
pg8000

# Utilities and API clients
python-dotenv==1.0.1
eodhd==1.0.32

# Additional dependencies
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.4.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mdurl==0.1.2
openai==1.59.8
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rich==13.9.4
scipy==1.15.1
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.8
websocket-client==1.8.0
websockets==14.1
